pyproject.toml
[project]
name = "ai-knowledge-base-rag"
version = "0.1.0"
description = "AI-powered knowledge base using RAG with completeness checks and enrichment suggestions"
requires-python = "==3.12.*"
dependencies = [
  "fastapi>=0.110,<0.116",
  "uvicorn>=0.27,<0.31",
  "python-multipart>=0.0.9,<0.1",
  "pydantic>=2.6,<2.9",
  "sqlalchemy>=2.0,<2.1",

  "faiss-cpu==1.13.2",
  "numpy==1.26.4",

  "python-docx==1.1.2",
  "pypdf==4.3.1",
  "jinja2==3.1.4",

  "openai==1.40.3",
  "httpx==0.27.2",
  "python-dotenv==1.0.1"
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0,<9.0"
]
